Extract QuestionnaireHostUiState and FhirJson into separate files
FhirJson moves to the util package so other repositories can share the
same encoding/decoding configuration.
