+++ ohs-player-reference-app/src/commonMain/kotlin/dev/ohs/player/reference/app/feature/questionnaire/QuestionnaireHostUiState.kt
@@ -1,0 +1,29 @@
+/*
+ * Copyright 2026 Open Health Stack Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.ohs.player.reference.app.feature.questionnaire
+
+sealed interface QuestionnaireHostUiState {
+  data object Loading : QuestionnaireHostUiState
+
+  data class Ready(val questionnaireJson: String, val title: String?) : QuestionnaireHostUiState
+
+  data class Submitting(val questionnaireJson: String, val title: String?) :
+    QuestionnaireHostUiState
+
+  data class Submitted(val result: QuestionnaireSubmissionResult) : QuestionnaireHostUiState
+
+  data class Error(val message: String) : QuestionnaireHostUiState
+}
