+++ ohs-player-reference-app/src/commonMain/kotlin/dev/ohs/player/reference/app/util/FhirJson.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2026 Open Health Stack Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *       http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package dev.ohs.player.reference.app.util
+
+import kotlinx.serialization.json.Json
+
+/** Single shared [Json] instance/configuration for encoding and decoding FHIR resources. */
+object FhirJson {
+  val instance: Json = Json {
+    prettyPrint = true
+    explicitNulls = false
+    encodeDefaults = false
+    ignoreUnknownKeys = true
+  }
+}
